+++ src/components/layout/Layout.tsx
@@ -1,5 +1,5 @@
 import { Box } from "@mui/material";
-import Header from "./Header";
+import Header from "./Header.tsx";
 import { Outlet } from "react-router-dom";
 
 export default function Layout() {
