+++ src/pages/Administration/pages/Authentication/components/PrivateRoute.tsx
@@ -1,5 +1,5 @@
 import {Navigate, Outlet} from "react-router-dom";
-import {useAuth} from "../../shared/hooks.ts";
+import {useAuth} from "../../../../../shared/hooks.ts";
 
 export default function PrivateRoute() {
   const {user} = useAuth()
