+++ src/pages/Results/components/StageHeader.tsx
