+++ src/pages/Results/pages/FootO/pages/Splits/SplitsStage.tsx
